requirements.txt
alabaster==0.7.8
amqp==1.4.9
anyjson==0.3.3
apiclient==1.0.3
appnope==0.1.0
attrs==16.0.0
awsebcli==3.8.7
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
behave==1.2.5
billiard==3.3.0.23
blessed==1.9.5
botocore==1.4.85
celery==3.1.23
cement==2.8.2
cffi==1.7.0
colorama==0.3.7
cryptography==1.4
cssselect==0.9.2
decorator==4.0.10
dj-database-url==0.4.1
Django==1.9.7
django-celery==3.1.17
django-cors-headers==1.1.0
django-extensions==1.6.1
django-fixture-magic==0.0.8
django-haystack==2.5.dev1
django-redis==4.6.0
django-redis-cache==1.6.3
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
elasticsearch==2.3.0
entrypoints==0.2.2
et-xmlfile==1.0.1
fuzzywuzzy==0.10.0
#gnureadline==6.3.3
gunicorn==19.6.0
idna==2.1
imagesize==0.7.1
ipykernel==4.3.1
ipyparallel==5.1.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==5.1.5
jdcal==1.2
Jinja2==2.8
jmespath==0.9.0
jsonfield==1.0.3
jsonschema==2.5.1
jupyter-client==4.3.0
jupyter-core==4.1.0
kombu==3.0.35
lxml==3.6.1
Markdown==2.6.6
MarkupSafe==0.23
mistune==0.7.2
nbconvert==4.2.0
nbformat==4.0.1
nose==1.3.7
notebook==4.2.1
numpy==1.11.2
openpyxl==2.3.5
pandas==0.19.1
parse==1.6.6
parse-type==0.3.4
parsel==1.0.3
path.py==8.2.1
pathspec==0.5.0
pbr==1.8.1
pexpect==4.1.0
pickleshare==0.7.2
pika==0.10.0
pike==0.1.1
psycopg2==2.6.1
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycrypto==2.6.1
pydash==3.4.3
PyDispatcher==2.0.5
pyexcel-io
Pygments==2.1.3
PyJWT==1.4.1
pymongo==3.3.0
pyOpenSSL==16.0.0
python-dateutil==2.6.0
python-Levenshtein==0.12.0
python-memcached==1.58
python-redis-lock==3.1.0
pytz==2016.4
PyYAML==3.12
pyzmq==15.2.0
qtconsole==4.2.1
queuelib==1.4.2
redis==2.10.5
#redis-cache==0.1.3
redis-structures==0.1.6
requests==2.10.0
requests-aws4auth==0.9
scikit-learn==0.18.1
scipy==0.18.1
Scrapy==1.1.1
scrapy-djangoitem==1.1.1
selenium==2.53.5
semantic-version==2.5.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
snowballstemmer==1.2.1
Sphinx==1.4.4
splinter==0.7.3
stevedore==1.12.0
tabulate==0.7.5
terminado==0.6
testpath==0.3
texttable==0.8.4
tornado==4.3
traitlets==4.2.1
Twisted==16.3.0
urllib3==1.16
virtualenv==15.0.1
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
#vital-tools==0.1.7
w3lib==1.15.0
wcwidth==0.1.7
websocket-client==0.40.0
whitenoise==3.2
Whoosh==2.7.4
widgetsnbextension==1.2.3
winappdbg==1.5
zope.interface==4.2.0
